Restructure netpol generate command to adhere more to kubectl semantics
