+++ roxctl/generate/generate.go
@@ -1,19 +1,19 @@
-package netpol
+package generate
 
 import (
 	"github.com/spf13/cobra"
 	"github.com/stackrox/rox/roxctl/common/environment"
-	"github.com/stackrox/rox/roxctl/netpol/generate"
+	"github.com/stackrox/rox/roxctl/generate/netpol"
 )
 
 // Command defines the collector command tree
 func Command(cliEnvironment environment.Environment) *cobra.Command {
 	c := &cobra.Command{
-		Use: "netpol",
+		Use: "generate",
 	}
 
 	c.AddCommand(
-		generate.Command(cliEnvironment),
+		netpol.Command(cliEnvironment),
 	)
 	return c
 }
