+++ roxctl/help/help.properties
@@ -181,11 +181,11 @@ helm.output.long =
 helm.derive-local-values.short = Derive local Helm values from cluster configuration.
 helm.derive-local-values.long =
 
-netpol.short = Commands related to recommending Network Policies.
-netpol.long =
+generate.short = Commands related to generating different resources.
+generate.long =
 
-netpol.generate.short = Recommend Network Policies based on deployment information.
-netpol.generate.long = Based on a given folder containing deployment YAMLs, will generate a list of recommended Network Policies.
+generate.netpol.short = Recommend Network Policies based on deployment information.
+generate.netpol.long = Based on a given folder containing deployment YAMLs, will generate a list of recommended Network Policies.
 
 version.short = Display the current roxctl version.
 version.long = Display the current roxctl version.
